requirements.txt
Flask>=3.0,<4.0
gunicorn>=21.2,<22.0
python-docx>=1.1.0,<2.0
